requirements.txt
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fake-useragent==1.5.1
Flask==3.0.2
Flask-Caching==2.1.0
Flask-WTF==1.2.1
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
outcome==1.3.0.post0
psutil==5.9.8
pycparser==2.21
PySocks==1.7.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
tabulate==0.9.0
tenacity==8.2.3
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
undetected-chromedriver==3.5.5
urllib3==2.2.1
watchdog==4.0.0
websockets==12.0
Werkzeug==3.0.1
wsproto==1.2.0
WTForms==3.1.2
zipp==3.18.1
